# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Soccer Api Documentation",
        "description": "Soccer API for doing CURD operation.",
        "contact": {
            "email": "b.sahoo@devon.nl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost.soccer.com/api",
            "description": "Soccer API Server"
        },
        {
            "url": "localhost.soccer.com/api"
        }
    ],
    "paths": {
...
        "/team/{teams}/players": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get list of player for a team",
                "description": "Returns list of players",
                "operationId": "getPlayerList",
                "parameters": [
                    {
                        "name": "teams",
                        "in": "path",
                        "description": "Teams id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlayerResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/player/{id}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get a player details",
                "description": "Returns list of players",
                "operationId": "getPlayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Player"
                ],
                "summary": "Delete existing player",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePlayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Player deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/player": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Store new player",
                "description": "To cerate a new player",
                "operationId": "storePlayer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "player/{id}": {
            "put": {
                "tags": [
                    "Player"
                ],
                "summary": "Update existing player",
                "description": "Update player data",
                "operationId": "updatePlayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Players updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Player": {
                "title": "Player",
                "description": "Player model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "title": "First Name",
                        "description": "First Name of the player",
                        "type": "string",
                        "example": "Lionel"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "Last Name of the player",
                        "type": "string",
                        "example": "Messi"
                    },
                    "image": {
                        "title": "Image",
                        "description": "Image file path of the player",
                        "type": "string",
                        "example": "/playre123625.jpg"
                    },
                    "team_id": {
                        "title": "Team Id",
                        "description": "Player's team ID",
                        "type": "integer",
                        "format": "int64",
                        "example": "3"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-02-15 15:55:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2022-02-15 15:55:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Player"
                }
            },
...
            "PlayerResource": {
                "title": "PlayerResource",
                "description": "Player resource",
                "properties": {
                    "data": {
                        "title": "Player Data",
                        "description": "Player Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Player"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PlayerResource"
                }
            },
...
            "StorePlayerRequest": {
                "title": "Store Player request",
                "description": "Store Player request body data",
                "required": [
                    "first_name",
                    "last_name",
                    "image",
                    "team_id"
                ],
                "properties": {
                    "first_name": {
                        "title": "First Name",
                        "description": "First Name of the player",
                        "type": "string",
                        "example": "Lionel"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "Last Name of the player",
                        "type": "string",
                        "example": "Messi"
                    },
                    "image": {
                        "title": "Image",
                        "description": "Image file of the player",
                        "example": "/playre123625.jpg"
                    },
                    "team": {
                        "title": "Team Id",
                        "description": "Player's team ID",
                        "type": "integer",
                        "format": "int64",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
            "UpdatePlayerRequest": {
                "title": "Update player request",
                "description": "Update player request body data",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "title": "First Name",
                        "description": "First Name of the player",
                        "type": "string",
                        "example": "Lionel"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "Last Name of the player",
                        "type": "string",
                        "example": "Messi"
                    },
                    "image": {
                        "title": "Image",
                        "description": "Image file of the player",
                        "example": "/playre123625.jpg"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "security": [
        []
    ]
}